sdks/aad-b2c/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aad"
version = "0.1.5"
authors = [
    {name="Marcel Claramunt", email="marcel@moveread.com"}
]
description = "Async Python SDK for validating Azure AD B2C access tokens"
dependencies = [
    "aiohttp", "pydantic", "pyjwt"
]
requires-python = ">=3.10"
readme = {file="README.md", content-type="text/markdown"}

[project.optional-dependencies]
fastapi = ["fastapi"]

[project.urls]
home = "https://github.com/moveread/azure-sdks/tree/main/sdks/aad-b2c"
repo = "https://github.com/moveread/azure-sdks"

# [project.optional-dependencies]
# test = [
#   "pytest < 5.0.0",
#   "pytest-cov[all]"
# ]